backend/requirements.txt
alembic==1.8.1
anyio==3.6.2
bcrypt==4.0.1
black==22.10.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.87.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.10.2
platformdirs==2.5.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.2
python-dotenv==0.21.0
python-jose==3.3.0
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
starlette==0.21.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
